@@ -6,8 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "eval": "node scripts/LLMJudgeEval.js --model deepseek-r1:7b --judge deepseek-r1:7b --dataset data/airport_eval_dataset.json"
+    "lint": "next lint"
   },
   "dependencies": {
     "@clerk/nextjs": "^4.29.7",
@@ -72,12 +71,14 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
     "autoprefixer": "^10.0.1",
-    "eslint": "^8",
+    "eslint": "^8.57.0",
     "eslint-config-next": "14.1.0",
     "postcss": "^8",
     "prisma": "^6.10.1",
     "tailwindcss": "^3.3.0",
-    "typescript": "^5"
+    "typescript": "^5.9.2"
   }
 }
